feat: see asset's information
