+++ fe/src/features/assets/create-new-asset/_components/index.ts
@@ -6,3 +6,7 @@ export * from './department-id-field'
 export * from './warrant-expiry-field'
 export * from './cost-field'
 export * from './owner-field'
+export * from './file-field'
+export * from './image-field'
+export * from './button-cancel'
+export * from './button-upload'
