+++ fe/src/features/assets/get-asset-detail/_components/asset-image.tsx
@@ -1,0 +1,17 @@
+import { Laptop } from 'lucide-react'
+import type { AssetsType } from '../../view-all-assets/model'
+
+export const AssetImage = ({ asset }: { asset: AssetsType }) => {
+  return asset.imageUpload ? (
+    <img
+      src={asset.imageUpload}
+      alt={asset.assetName}
+      className='max-h-[300px] rounded-md object-contain'
+    />
+  ) : (
+    <div className='flex h-[300px] w-full flex-col items-center justify-center rounded-md border border-dashed'>
+      <Laptop className='text-muted-foreground h-10 w-10' />
+      <p className='text-muted-foreground mt-2 text-sm'>No image available</p>
+    </div>
+  )
+}
