+++ fe/src/features/assets/get-asset-detail/_components/index.ts
@@ -1,0 +1,5 @@
+export * from './asset-badge'
+export * from './no-asset'
+export * from './asset-information'
+export * from './asset-image'
+export * from './asset-file'
