+++ fe/src/features/assets/get-asset-detail/get-asset-detail.tsx
@@ -5,36 +5,22 @@ import { useEffect, useTransition } from 'react'
 import { getAssetInformation } from '../api'
 import { useState } from 'react'
 import { toast } from 'sonner'
-import { format } from 'date-fns'
 import {
   Card,
   CardContent,
   CardDescription,
-  CardFooter,
   CardHeader,
   CardTitle,
   Button,
-  Badge,
   Tabs,
   TabsContent,
   TabsList,
   TabsTrigger,
 } from '@/components/ui'
 
-import {
-  Laptop,
-  Calendar,
-  DollarSign,
-  Tag,
-  Building,
-  User,
-  FileText,
-  Clock,
-  ArrowLeft,
-  Pencil,
-  Loader2,
-} from 'lucide-react'
-import { cn } from '@/lib'
+import { ArrowLeft, Pencil, Loader2 } from 'lucide-react'
+
+import { AssetBadge, AssetFile, AssetImage, AssetInformation, NoAsset } from './_components'
 
 const GetAssetDetail = () => {
   const { id } = useParams()
@@ -66,26 +52,7 @@ const GetAssetDetail = () => {
   }
 
   if (!asset) {
-    return (
-      <div className='container mx-auto px-4 py-8'>
-        <Card className='border-destructive'>
-          <CardHeader>
-            <CardTitle className='text-destructive'>Asset Not Found</CardTitle>
-          </CardHeader>
-          <CardContent>
-            <p>We couldn't find the asset with ID: {id}</p>
-          </CardContent>
-          <CardFooter>
-            <Link to='/assets'>
-              <Button>
-                <ArrowLeft className='mr-2 h-4 w-4' />
-                Back to Assets
-              </Button>
-            </Link>
-          </CardFooter>
-        </Card>
-      </div>
-    )
+    return <NoAsset id={id || ''} />
   }
 
   return (
@@ -101,19 +68,7 @@ const GetAssetDetail = () => {
             </Button>
           </Link>
           <h1 className='text-3xl font-semibold'>{asset.assetName}</h1>
-          <Badge
-            variant='outline'
-            className={cn(
-              'ml-4 flex items-center gap-1',
-              asset.status === 'New' && 'border-green-200 bg-green-100 text-green-800',
-              asset.status === 'In Use' && 'border-blue-200 bg-blue-100 text-blue-800',
-              asset.status === 'Under Maintenance' && 'border-amber-200 bg-amber-100 text-amber-800',
-              asset.status === 'Retired' && 'border-slate-200 bg-slate-100 text-slate-800',
-              asset.status === 'Disposed' && 'border-gray-200 bg-gray-100 text-gray-800'
-            )}
-          >
-            {asset.status}
-          </Badge>
+          <AssetBadge asset={asset} />
         </div>
         <Link to={`/assets/edit/${id}`}>
           <Button variant='outline'>
@@ -131,63 +86,7 @@ const GetAssetDetail = () => {
               <CardDescription>Details about the hardware asset</CardDescription>
             </CardHeader>
             <CardContent className='space-y-6'>
-              <div className='grid grid-cols-1 gap-y-6 sm:grid-cols-2'>
-                <div className='space-y-1'>
-                  <h3 className='text-muted-foreground text-sm font-medium'>Asset Type</h3>
-                  <p className='flex items-center font-medium'>
-                    <Laptop className='text-primary mr-2 h-4 w-4' />
-                    {asset.category?.categoryName}
-                  </p>
-                </div>
-
-                <div className='space-y-1'>
-                  <h3 className='text-muted-foreground text-sm font-medium'>Serial Number</h3>
-                  <p className='flex items-center font-medium'>
-                    <Tag className='text-primary mr-2 h-4 w-4' />
-                    {asset.serialNumber}
-                  </p>
-                </div>
-
-                <div className='space-y-1'>
-                  <h3 className='text-muted-foreground text-sm font-medium'>Purchase Date</h3>
-                  <p className='flex items-center font-medium'>
-                    <Calendar className='text-primary mr-2 h-4 w-4' />
-                    {format(new Date(asset.purchaseDate), 'PPP')}
-                  </p>
-                </div>
-
-                <div className='space-y-1'>
-                  <h3 className='text-muted-foreground text-sm font-medium'>Warranty Expiry</h3>
-                  <p className='flex items-center font-medium'>
-                    <Clock className='text-primary mr-2 h-4 w-4' />
-                    {format(new Date(asset.warrantExpiry), 'PPP')}
-                  </p>
-                </div>
-
-                <div className='space-y-1'>
-                  <h3 className='text-muted-foreground text-sm font-medium'>Department</h3>
-                  <p className='flex items-center font-medium'>
-                    <Building className='text-primary mr-2 h-4 w-4' />
-                    {asset.department?.departmentName}
-                  </p>
-                </div>
-
-                <div className='space-y-1'>
-                  <h3 className='text-muted-foreground text-sm font-medium'>Cost</h3>
-                  <p className='flex items-center font-medium'>
-                    <DollarSign className='text-primary mr-2 h-4 w-4' />
-                    {asset.cost}
-                  </p>
-                </div>
-
-                <div className='space-y-1'>
-                  <h3 className='text-muted-foreground text-sm font-medium'>Owner</h3>
-                  <p className='flex items-center font-medium'>
-                    <User className='text-primary mr-2 h-4 w-4' />
-                    {asset.owner?.firstName + ' ' + asset.owner?.lastName}
-                  </p>
-                </div>
-              </div>{' '}
+              <AssetInformation asset={asset} />
             </CardContent>
           </Card>
         </div>
@@ -215,18 +114,7 @@ const GetAssetDetail = () => {
             >
               <Card>
                 <CardContent className='flex items-center justify-center p-4'>
-                  {asset.imageUpload ? (
-                    <img
-                      src={asset.imageUpload}
-                      alt={asset.assetName}
-                      className='max-h-[300px] rounded-md object-contain'
-                    />
-                  ) : (
-                    <div className='flex h-[300px] w-full flex-col items-center justify-center rounded-md border border-dashed'>
-                      <Laptop className='text-muted-foreground h-10 w-10' />
-                      <p className='text-muted-foreground mt-2 text-sm'>No image available</p>
-                    </div>
-                  )}
+                  <AssetImage asset={asset} />
                 </CardContent>
               </Card>
             </TabsContent>
@@ -237,30 +125,7 @@ const GetAssetDetail = () => {
             >
               <Card>
                 <CardContent className='p-4'>
-                  {asset.fileAttachment ? (
-                    <div className='flex items-center justify-between rounded-md border p-3'>
-                      <div className='flex items-center'>
-                        <FileText className='text-primary mr-2 h-5 w-5' />
-                        <span>Asset Document</span>
-                      </div>
-                      <Link
-                        to={asset.fileAttachment}
-                        download={true}
-                      >
-                        <Button
-                          variant='outline'
-                          size='sm'
-                        >
-                          View
-                        </Button>
-                      </Link>
-                    </div>
-                  ) : (
-                    <div className='flex h-[100px] w-full flex-col items-center justify-center rounded-md border border-dashed'>
-                      <FileText className='text-muted-foreground h-8 w-8' />
-                      <p className='text-muted-foreground mt-2 text-sm'>No documents available</p>
-                    </div>
-                  )}
+                  <AssetFile asset={asset} />
                 </CardContent>
               </Card>
             </TabsContent>
