+++ fe/src/features/auth/forget-password-form/_components/index.ts
