+++ fe/src/features/auth/forget-password-form/_components/sign-up-success.tsx
